CloudFunctions/traffic_function/requirements.txt
google-cloud-firestore
google-maps-routing
google-auth
python-dotenv
pytz
